Restructing project
