+++ data/ie_inflation_cpi.csv
@@ -545,3 +545,8 @@ month,observation
 2022-02,5.6
 2022-03,6.7
 2022-04,7.0
+2022-05,7.8
+2022-06,9.1
+2022-07,9.1
+2022-08,8.7
+2022-09,8.2
