+++ pyproject.toml
@@ -1,0 +1,25 @@
+[tool.poetry]
+name = "nso-stats-fetcher"
+version = "0.1.0"
+description = "National Statistical Offices Statistics Fetcher"
+packages = [{include = "nso_stats_fetcher"}]
+authors = []
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+pandas = "^1.5.1"
+requests = "^2.28.1"
+openpyxl = "^3.0.10"
+tabula-py = "^2.5.1"
+xlrd = "^2.0.1"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
