+++ src/nsofetch/analysis.py
