+++ src/nsofetch/fetch_ar.py
