+++ src/nsofetch/fetch_ph.py
@@ -1,0 +1,134 @@
+import pandas
+import numpy
+
+import filepaths
+import utils
+
+
+def fetch_ph_inflation_cpi():
+    stats_metadata = utils.read_stats_metadata()
+
+    # Couldn't get a machine readable file, so had to directly copy the values from this page:
+    # https://psa.gov.ph/statistics/survey/price/summary-inflation-report-consumer-price-index-2018100-may-2022
+
+    times = [
+        "2018-01",
+        "2018-02",
+        "2018-03",
+        "2018-04",
+        "2018-05",
+        "2018-06",
+        "2018-07",
+        "2018-08",
+        "2018-09",
+        "2018-10",
+        "2018-11",
+        "2018-12",
+        "2019-01",
+        "2019-02",
+        "2019-03",
+        "2019-04",
+        "2019-05",
+        "2019-06",
+        "2019-07",
+        "2019-08",
+        "2019-09",
+        "2019-10",
+        "2019-11",
+        "2019-12",
+        "2020-01",
+        "2020-02",
+        "2020-03",
+        "2020-04",
+        "2020-05",
+        "2020-06",
+        "2020-07",
+        "2020-08",
+        "2020-09",
+        "2020-10",
+        "2020-11",
+        "2020-12",
+        "2021-01",
+        "2021-02",
+        "2021-03",
+        "2021-04",
+        "2021-05",
+        "2021-06",
+        "2021-07",
+        "2021-08",
+        "2021-09",
+        "2021-10",
+        "2021-11",
+        "2021-12",
+        "2022-01",
+        "2022-02",
+        "2022-03",
+        "2022-04",
+        "2022-05",
+    ]
+
+    observations = [
+        3.4,
+        3.7,
+        4.3,
+        4.3,
+        4.6,
+        5.0,
+        5.8,
+        6.6,
+        6.9,
+        6.9,
+        6.1,
+        5.2,
+        4.4,
+        3.8,
+        3.4,
+        3.2,
+        3.2,
+        2.7,
+        2.2,
+        1.4,
+        0.5,
+        0.6,
+        1.2,
+        2.4,
+        3.0,
+        2.5,
+        2.2,
+        1.8,
+        1.6,
+        2.3,
+        2.4,
+        2.2,
+        2.2,
+        2.3,
+        3.0,
+        3.3,
+        3.7,
+        4.2,
+        4.1,
+        4.1,
+        4.1,
+        3.7,
+        3.7,
+        4.4,
+        4.2,
+        4.0,
+        3.7,
+        3.1,
+        3.0,
+        3.0,
+        4.0,
+        4.9,
+        5.4,
+    ]
+
+    output_df = pandas.DataFrame({"month": times, "observation": observations})
+    output_filepath = (
+        filepaths.DATA_DIR / stats_metadata["PH"]["inflation"]["CPI"]["filename"]
+    )
+    output_df.to_csv(output_filepath, index=False)
+
+
+if __name__ == "__main__":
+    fetch_ph_inflation_cpi()
